refactor(media): promote WSAudioPipe with inner AudioInput/Output
